refactor: Extract useSavedBuildings custom hook
